Added new attribute and fixed issues with reorderable list.
Fixed header for reorderable list so that it does not render on top of first element.
Fixed element header for reorderable list so that it can be customized.
Added new MultiConditionalFieldAttribute to evaluate multiple control fields.
Updated ConditionalFieldAttribute and MultiConditionalFieldAttribute to enable 'Clear Otherwise' to blank value if the field is not shown in the editor.
Updated version number in project json.